requirements.txt
jira==3.1.1
pick==1.2.0
python-dotenv==0.19.2
